refactor: simplify ActivityHeatmap component by utilizing custom hooks for grid and scroll management
