+++ src/renderer/components/states/maxExpand/components/cli/components/ActivityHeatmap.tsx
@@ -24,10 +24,12 @@
  * @author 鸡哥
  */
 
-import { useEffect, useMemo, useRef, useState, type ReactElement } from 'react';
+import { useState, type ReactElement } from 'react';
 import { useTranslation } from 'react-i18next';
 import type { ActivityHeatmapProps, HeatmapMetric } from '../types/types';
-import { buildHeatmapMonths, HEATMAP_MONTH_KEYS } from '../utils/heatmapGrid';
+import { HEATMAP_MONTH_KEYS } from '../utils/heatmapGrid';
+import { useHeatmapGrid } from '../hooks/useHeatmapGrid';
+import { useHeatmapScroll } from '../hooks/useHeatmapScroll';
 
 /**
  * Claude Code 活动热力图组件
@@ -37,40 +39,8 @@ import { buildHeatmapMonths, HEATMAP_MONTH_KEYS } from '../utils/heatmapGrid';
 export function ActivityHeatmap({ heatmap, compact = false, visible = true }: ActivityHeatmapProps): ReactElement {
   const { t } = useTranslation();
   const [metric, setMetric] = useState<HeatmapMetric>('session');
-  const scrollRef = useRef<HTMLDivElement>(null);
-  const todayRef = useRef<HTMLSpanElement>(null);
-
-  const grid = useMemo(() => {
-    const totals = { session: 0, tool: 0, prompt: 0 };
-    const dayCounts = new Map<string, number>();
-    Object.entries(heatmap).forEach(([key, counts]) => {
-      totals.session += counts.session;
-      totals.tool += counts.tool;
-      totals.prompt += counts.prompt;
-      const value = counts[metric];
-      if (value > 0) dayCounts.set(key, value);
-    });
-    const { months, max } = buildHeatmapMonths((key) => dayCounts.get(key) ?? 0);
-    return { months, max, totals };
-  }, [heatmap, metric]);
-
-  /** 把数量映射到 0-4 档色阶强度 */
-  const levelOf = (count: number): number => {
-    if (count <= 0) return 0;
-    return Math.min(4, Math.ceil((count / grid.max) * 4));
-  };
-
-  // 挂载、可见性切换或指标切换时把今日滚动到水平居中
-  useEffect(() => {
-    if (!visible) return;
-    const id = requestAnimationFrame(() => {
-      const scroller = scrollRef.current;
-      const today = todayRef.current;
-      if (!scroller || !today) return;
-      scroller.scrollLeft = today.offsetLeft - scroller.clientWidth / 2 + today.offsetWidth / 2;
-    });
-    return () => cancelAnimationFrame(id);
-  }, [visible, metric]);
+  const grid = useHeatmapGrid(heatmap, metric);
+  const { scrollRef, todayRef } = useHeatmapScroll(visible, metric);
 
   const variantClass = compact ? ' cli-tab-heatmap--compact' : '';
 
@@ -114,7 +84,7 @@ export function ActivityHeatmap({ heatmap, compact = false, visible = true }: Ac
                       <span
                         key={cell.key}
                         ref={cell.isToday ? todayRef : undefined}
-                        className={`cli-tab-heatmap-cell${cell.future ? ' cli-tab-heatmap-cell--future' : ` level-${levelOf(cell.count)}`}${cell.isToday ? ' cli-tab-heatmap-cell--today' : ''}`}
+                        className={`cli-tab-heatmap-cell${cell.future ? ' cli-tab-heatmap-cell--future' : ` level-${grid.levelOf(cell.count)}`}${cell.isToday ? ' cli-tab-heatmap-cell--today' : ''}`}
                         style={{ gridColumnStart: Math.floor(pos / 7) + 1, gridRowStart: (pos % 7) + 1 }}
                         title={cell.future ? '' : `${cell.label}: ${cell.count}`}
                       />
